refactor: new svm module `account_saver` (#2167)
