+++ svm/src/lib.rs
@@ -4,6 +4,7 @@
 pub mod account_loader;
 pub mod account_overrides;
 pub mod account_rent_state;
+pub mod account_saver;
 pub mod message_processor;
 pub mod nonce_info;
 pub mod program_loader;
